refactor: reorganizing components
